requirements.txt
streamlit==1.46.0          # UI
pandas==2.3.0              # CSV 처리
numpy==1.26.4
scipy==1.11.4
tqdm==4.67.1
python-dotenv==1.1.1       # .env
requests==2.32.3           # 스크래퍼
beautifulsoup4==4.13.0
httpx==0.28.0

# 오디오 분석
librosa==0.10.1
soundfile==0.12.1
mutagen==1.47.0
soxr==0.5.0.post1

# 기능성
rapidfuzz==3.13.0
lxml==5.4.0
scikit-learn==1.5.0

# 빌드
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
